lab6/exercises/trainfun/requirements.txt
pandas
keras
tensorflow
numpy
h5py
google-cloud-storage
flask